Changes EntityManager Factory to EntityManagerFactory Factory
It takes the most time to create a new EMF and it doesn't seem to
take any time to create an EM from that EMF. As such, it seems
better to have a singleton only for the EMF and create EMs from it
as needed.

I am also closing every EM when the transaction is done.