Allows user to create/delete/edit Students & Accounts on TreeView
Using two TreeItems, one for Students and one for Accounts, users
can now view Accounts and Students in the TreeView. Both TreeItems
have their own ContextMenus to perform actions. They can be edited
too due to the TreeCellFactory.

I think there may be a better way to perform the actions and update
the TreeView and database. Possibly with the observer pattern.