Styles the GUI
There is one small problem: the save button on the notes view is a
different shape than the save button on the other views. This is
because of the BorderPane layout. The notes view does not use the
left borderpane, so the button on the bottom is a different shape
than the other buttons on other views.